Spread all the files out to make it easier to read
