+++ client/commands/all.py
@@ -1,0 +1,2 @@
+def cmd_all(ws, user_id, privkey_pem, text):
+    print("all")
